feat: add support for sending messages with Ctrl+Enter
This commit modifies the `initInputContainer` function to add support for sending messages with Ctrl+Enter. If the user presses Ctrl+Enter, the message is sent without adding a new line. If the user presses Enter without Ctrl or Shift, a new line is added to the input field. The commit also removes unnecessary whitespace.