feat: Add context menu icons
Add icons to the context menu items based on their names. The function contextMenuIcon() was added to the code to handle this. The function returns an icon based on the name of the context menu item.

The following icons were added: IconMouseRightClick, IconBook, IconGitBranch, IconGitBranchChecked, and IconShellCommand.

The code was also refactored to use the contextMenuIcon() function instead of the IconMessagePlus component.