feat: Add new avatar image and adjust chat panel layout
Added a new avatar image and adjusted the layout of the chat panel to accommodate the new image. The new avatar image is now displayed for user messages. The layout of the chat panel has been adjusted to ensure that the new image is displayed correctly. The width of the chat panel has been set to the width of the container. The padding and margin of the chat panel have been adjusted to ensure that the new image is displayed correctly.