ci: Add CircleCI configuration for automated build and publish
This commit introduces a CircleCI configuration that enables automatic building of the VSCode extension on every main branch update. It also sets up automatic publishing to the VSCode Marketplace whenever a new tag is created on the main branch. This workflow significantly improves the CI/CD process of the extension development.

Signed-off-by: Daniel Hu <tao.hu@merico.dev>